requirements/requirements.txt
#Core

Django==1.7
djangorestframework==2.4.3
South==1.0
django_compressor==1.4
django-static-precompiler==0.7

#Deployment

Fabric==1.10.0

#Tools

django-debug-toolbar==1.2.1
django-rest-framework-docs==0.1.7
ipython==2.3.0
Sphinx==1.2.3